+++ soh/CMakeLists.txt
@@ -162,10 +162,6 @@ set(Header_Files__soh__Enhancements__controls
 )
 source_group("Header Files\\soh\\Enhancements\\controls" FILES ${Header_Files__soh__Enhancements__controls})
 
-set(Header_Files__soh__Enhancements__menu
-    "soh/Enhancements/menu/GameMenuBar.hpp"
-)
-source_group("Header Files\\soh\\Enhancements\\menu" FILES ${Header_Files__soh__Enhancements__menu})
 
 set(Header_Files__soh__Enhancements__cosmetics
     "soh/Enhancements/cosmetics/CosmeticsEditor.h"
@@ -256,6 +252,10 @@ set(Source_Files__soh
     "soh/z_message_OTR.cpp"
     "soh/z_play_otr.cpp"
     "soh/z_scene_otr.cpp"
+    "soh/GameMenuBar.hpp"
+    "soh/GameMenuBar.cpp"
+    "soh/UIWidgets.hpp"
+    "soh/UIWidgets.cpp"
 )
 source_group("Source Files\\soh" FILES ${Source_Files__soh})
 
@@ -271,11 +271,6 @@ set(Source_Files__soh__Enhancements__controls
     "soh/Enhancements/controls/GameControlEditor.cpp"
 )
 source_group("Source Files\\soh\\Enhancements\\controls" FILES ${Source_Files__soh__Enhancements__controls})
-
-set(Source_Files__soh__Enhancements__menu
-    "soh/Enhancements/menu/GameMenuBar.cpp"
-)
-source_group("Source Files\\soh\\Enhancements\\menu" FILES ${Source_Files__soh__Enhancements__menu})
 
 set(Source_Files__soh__Enhancements__cosmetics
     "soh/Enhancements/cosmetics/CosmeticsEditor.cpp"
@@ -1578,7 +1573,6 @@ set(ALL_FILES
     ${Header_Files__include}
     ${Header_Files__soh__Enhancements}
     ${Header_Files__soh__Enhancements__controls}
-    ${Header_Files__soh__Enhancements__menu}
     ${Header_Files__soh__Enhancements__cosmetics}
     ${Header_Files__soh__Enhancements__debugger}
     ${Header_Files__soh__Enhancements__randomizer}
@@ -1588,7 +1582,6 @@ set(ALL_FILES
     ${Source_Files__soh}
     ${Source_Files__soh__Enhancements}
     ${Source_Files__soh__Enhancements__controls}
-    ${Source_Files__soh__Enhancements__menu}
     ${Source_Files__soh__Enhancements__cosmetics}
     ${Source_Files__soh__Enhancements__debugger}
     ${Source_Files__soh__Enhancements__randomizer}
